requirements.txt
aioamqp==0.14.0
aioredis==1.3.1
asgiref==3.3.1
async-timeout==3.0.1
attrs==20.3.0
autobahn==20.7.1
Automat==20.2.0
boto3==1.16.34
botocore==1.19.34
certifi==2020.12.5
cffi==1.14.4
channels==3.0.2
channels-redis==3.2.0
chardet==3.0.4
constantly==15.1.0
cryptography==3.3.1
daphne==3.0.1
Django==3.1.4
djangorestframework==3.12.2
docutils==0.16
hiredis==1.1.0
hyperlink==20.0.1
idna==2.10
incremental==17.5.0
itsdangerous==1.1.0
jmespath==0.10.0
msgpack==1.0.1
numpy==1.19.4
pamqp==3.0.1
psycopg2==2.8.6
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
PyHamcrest==2.0.2
PyJWT==1.7.1
pyOpenSSL==20.0.0
python-dateutil==2.8.1
pytz==2020.4
requests==2.25.0
s3transfer==0.3.3
service-identity==18.1.0
six==1.15.0
sqlparse==0.4.1
Twisted==20.3.0
txaio==20.4.1
urllib3==1.26.2
whitenoise==5.2.0
xlwt==1.3.0
zope.interface==5.2.0
